chore: refactor notifier package
Split the notifier package into smaller source files.
